Refactorings:
- Organize model by component
- Let client methods return the domain object (as async) instead of being wrapped in `EventLoopFuture`.

This uses `get` which puts some restrictions on supported macOS versions, but all recent versions are still supported.
