+++ Sources/ConcordiumSwiftSDK/Client.swift
@@ -1,8 +1,8 @@
 import Foundation
 
 import GRPC
-import NIOPosix
 import NIOCore
+import NIOPosix
 
 class Client {
     let grpc: Concordium_V2_QueriesNIOClient
@@ -11,38 +11,23 @@ class Client {
         grpc = Concordium_V2_QueriesNIOClient(channel: channel)
     }
 
-    private func getBlockHashInput(blockHash: BlockHash?) -> Concordium_V2_BlockHashInput {
-        if let blockHash {
-            var h = Concordium_V2_BlockHash()
-            h.value = blockHash
-            var b = Concordium_V2_BlockHashInput()
-            b.given = h
-            return b
-        } else {
-            var b = Concordium_V2_BlockHashInput()
-            b.lastFinal = Concordium_V2_Empty()
-            return b
-        }
+    func getCryptographicParameters(at block: BlockIdentifier) async throws -> CryptographicParameters {
+        var req = block.toGrpcType()
+        let res = try await grpc.getCryptographicParameters(req).response.get()
+        return CryptographicParameters(
+                onChainCommitmentKey: res.onChainCommitmentKey.hexadecimalString(),
+                bulletproofGenerators: res.bulletproofGenerators.hexadecimalString(),
+                genesisString: res.genesisString
+        )
     }
 
-    func getCryptographicParameters(at block: BlockIdentifier) async -> EventLoopFuture<CryptographicParameters> {
-        return grpc
-                .getCryptographicParameters(block.toGrpcType())
-                .response
-                .map({ v in
-                    CryptographicParameters(
-                            onChainCommitmentKey: v.onChainCommitmentKey.hexadecimalString(),
-                            bulletproofGenerators: v.bulletproofGenerators.hexadecimalString(),
-                            genesisString: v.genesisString
-                    )
-                })
-    }
-
-    func getNextAccountSequenceNumber(forAddress address: AccountAddress) -> EventLoopFuture<NextAccountSequenceNumber?> {
-        var grpcAddress = Concordium_V2_AccountAddress()
-        grpcAddress.value = address
-        return grpc.getNextAccountSequenceNumber(grpcAddress).response.map({ res in
-            NextAccountSequenceNumber(sequenceNumber: res.sequenceNumber.value, allFinal: res.allFinal)
-        })
+    func getNextAccountSequenceNumber(of address: AccountAddress) async throws -> NextAccountSequenceNumber {
+        var req = Concordium_V2_AccountAddress()
+        req.value = address
+        let res = try await grpc.getNextAccountSequenceNumber(req).response.get()
+        return NextAccountSequenceNumber(
+                sequenceNumber: res.sequenceNumber.value,
+                allFinal: res.allFinal
+        )
     }
 }
