+++ Sources/ConcordiumSwiftSDK/Model/Chain.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct CryptographicParameters {
+    let onChainCommitmentKey: String
+    let bulletproofGenerators: String
+    let genesisString: String
+}
